fast-agent-bot/requirements.txt
fast-agent-mcp>=0.1.0
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
